+++ resources/js/components/tools/tools.js
@@ -1,5 +1,5 @@
-import React from "react";
 import alertify from 'alertifyjs';
+import  { reloadTableData } from "../table/table.js";
 
 /**
  * guardar los datos en la funcion store de laravel
@@ -7,6 +7,9 @@ import alertify from 'alertifyjs';
  * @param dataForm = son los datos que se van a enviar al servidor
  *
  */
+let token = document
+    .querySelector("meta[name='csrf-token']")
+    .getAttribute("content");
 
 export const saveDataForm = async (urlSave, dataForm, elementId, messageSend = null) => {
     var request = "";
@@ -102,33 +105,39 @@ export const getById = async (url, id) => {
         );
     return request;
 };
-export const deletedElement = async (url, id) => {
-    console.log("getById");
-    let token = document
-        .querySelector("meta[name='csrf-token']")
-        .getAttribute("content");
-    const request = await fetch(url + "/" + id, {
-        method: "DELETE",
-        headers: {
-            "X-CSRF-TOKEN": token,
-            Accept: "application/json",
-            "Content-Type": "application/json"
-        }
-    })
+export const deletedElement =  (url, id, nombre= "") => {
+    alertify.confirm('Eliminar item', 'Esta seguro que desea eliminar ' + nombre, async function(){
+         var request = await fetch(url + "/" + id, {
+            method: "DELETE",
+            headers: {
+                "X-CSRF-TOKEN": token,
+                Accept: "application/json",
+                "Content-Type": "application/json"
+            }
+        })
         .then(res => res.json())
         .then(
-            result => {
-                return result;
-            },
-            error => {
-                console.log(error);
-            }
-        );
-    return request;
+                result => {
+                    if( result.status )
+                    {
+                        alertify.success(result.message);
+                        reloadTableData();
+                    }
+                    return result;
+                },
+                error => {
+
+                    console.log(error);
+                }
+            );
+         return request;
+    },
+     function(){
+        alertify.error('Cancelado');
+        });
+
 };
-let token = document
-    .querySelector("meta[name='csrf-token']")
-    .getAttribute("content");
+
 export const getAllByClass = async (url, dataForm) => {
     const request = await fetch(url, {
         method: "POST",
